Added getUserInfo and Oauth functionality.
The Oauth does not quite work with a real application as I'm not sure how the authorize request is sent to Salesforce (not seeing it in sample app with http and sf logging).  Will need to step through the connector code next week, but I'm overdue in committing my changes.